fix TimeBeforeEnd (need cast)

+++ Game/Abnormality/Abnormality.cs
@@ -40,7 +40,7 @@ namespace Tera.Game.Abnormality
 
         public long FirstHit { get; private set; }
 
-        public long TimeBeforeEnd => Duration != 0 ? FirstHit - DateTime.UtcNow.Ticks + Duration*10000000 : long.MaxValue;
+        public long TimeBeforeEnd => Duration != 0 ? FirstHit - DateTime.UtcNow.Ticks + (long)Duration*10000000 : long.MaxValue;
         public long TimeBeforeApply => DateTime.UtcNow.Ticks - LastApply - HotDot.Tick*10000000;
 
         public void Apply(int amount, bool critical, bool isHp, long time)
